refactor: Move a function to helpers
