+++ src/ducks/transactions/helpers.js
@@ -1,4 +1,35 @@
+import find from 'lodash/find'
+import findLast from 'lodash/findLast'
+
 export const getLabel = transaction =>
   transaction.label.toLowerCase().replace(/(?:^|\s)\S/g, a => a.toUpperCase())
 
 export const getDate = transaction => transaction.date.slice(0, 10)
+
+// Performs successive `find`s until one of the find functions returns
+// a result
+const multiFind = (arr, findFns) => {
+  for (let findFn of findFns) {
+    const res = findFn(arr)
+    if (res) {
+      return res
+    }
+  }
+  return null
+}
+
+/**
+ * Returns the first month having operations, closest to the given month.
+ *
+ * To know if we have to search in the past or the in the future, we check
+ * if the chosen month is before or after the current month.
+ */
+export const findNearestMonth = (chosenMonth, currentMonth, availableMonths) => {
+  const findBeforeChosenMonth = months => findLast(months, x => x < chosenMonth)
+  const findAfterChosenMonth = months => find(months, x => x > chosenMonth)
+  const findFns =
+    chosenMonth < currentMonth
+      ? [findBeforeChosenMonth, findAfterChosenMonth]
+      : [findAfterChosenMonth, findBeforeChosenMonth]
+  return multiFind(availableMonths, findFns)
+}
